+++ changelog.md
@@ -6,6 +6,8 @@
 This version is not released yet and is under active development.
 ```
 
+- Add test to search in Pygments' test data for REPL/terminal-like lexers, as candidates for ANSI-coloring.
+
 ## {gh}`3.6.0 (2022-12-28) <compare/v3.5.0...v3.6.0>`
 
 - Add new constants to group platforms by family.